hostelAdminV2/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pymongo
certifi
passlib[bcrypt]
python-jose[cryptography]
python-multipart
pydantic[email]
razorpay
